requirements.txt
sentence-transformers
rank_bm25
faiss-cpu
pandas
numpy
python-dotenv
requests
